add match_events field to MatchResult, add jsonfield dependency
Took 45 minutes

+++ game_engine/views.py
@@ -46,32 +46,36 @@ class MatchViewSet(viewsets.ModelViewSet):
     # noinspection PyUnusedLocal,PyShadowingBuiltins
     @action(methods=["POST"], detail=True, permission_classes=[])
     def report_match(self, request, pk=None, format=None):  # todo: turn into serializer
+        # print(request.data)
         if "winners" in request.data and isinstance(request.data["winners"], list):
             match = Match.objects.get(pk=pk)
             match_players = match.players
             winners = request.data["winners"]
             losers = set(match_players).difference(set(winners))
             if set(winners).issubset(set(match_players)):
-                match_result = MatchResult()
-                match_result.time_started = match.allocated
-                match_result.players = Match.objects.get(pk=pk).players
-                match_result.winners = request.data["winners"]
-                match_result.save()
+                if "match_history" in request.data and isinstance(request.data["match_history"], list):
+                    match_result = MatchResult()
+                    match_result.time_started = match.allocated
+                    match_result.players = Match.objects.get(pk=pk).players
+                    match_result.winners = request.data["winners"]
+                    match_result.match_events = request.data["match_history"]
+                    match_result.save()
 
-                match.delete()
+                    match.delete()
 
-                for player in match_players:
-                    up_instance, created = UserPerformance.objects.get_or_create(user=User.objects.get(pk=player))
-                    up_instance.games_played += 1
-                    up_instance.save()
-                # UserPerformance.objects.filter(user__pk__in=match_players).update(games_played=F('games_played')+1)
-                UserPerformance.objects.filter(user__pk__in=winners).update(mmr=F('mmr') + 100)
-                UserPerformance.objects.filter(user__pk__in=losers).update(mmr=F('mmr') - 100)
+                    for player in match_players:
+                        up_instance, created = UserPerformance.objects.get_or_create(user=User.objects.get(pk=player))
+                        up_instance.games_played += 1
+                        up_instance.save()
+                    # UserPerformance.objects.filter(user__pk__in=match_players).update(games_played=F('games_played')+1)
+                    UserPerformance.objects.filter(user__pk__in=winners).update(mmr=F('mmr') + 100)
+                    UserPerformance.objects.filter(user__pk__in=losers).update(mmr=F('mmr') - 100)
 
-                UserCode.objects.filter(user__pk__in=match_players).update(is_in_game=False)
-                # todo: implement actual MMR calculation
-                return Response(status=status.HTTP_201_CREATED)
-
+                    UserCode.objects.filter(user__pk__in=match_players).update(is_in_game=False)
+                    # todo: implement actual MMR calculation
+                    return Response(status=status.HTTP_201_CREATED)
+                return Response({"ok": False, "message": "Missing match history"},
+                                status=status.HTTP_400_BAD_REQUEST)
             return Response({"ok": False, "message": "One or more winner not part of match"},
                             status=status.HTTP_400_BAD_REQUEST)
         return Response({"ok": False, "message": "No winners provided"},
